requirements.txt
pyyaml
flask
pymysql
requests